Use cache instead of filecache
I misunderstood what filecache does - I thought it would cache results
to a file, so the cache could persist through program restarts. Instead,
it simply uses the file names as a cache expiry mechanism: when the file
that is the cache key is modified, the corresponding cache entry
expires.

The cache package uses a simpler time-based expiration, which is what I
actually wanted.